+++ RELEASENOTES.md
@@ -23,6 +23,7 @@
 - Fix lite amp goggles/render.fullbright not increasing the light level when using palette color with true color overlays disabled.
 - Fix sprites being generated and duplicated at runtime in true color.
 - Fix font coloring that was rendering dark in menus (fixes Antaresian Reliquary colored font rendering).
+- Fix sky texture not working when it's not in the texture namespace.
 
 ## Misc:
 - Refactor of old Status Bar renderer to data-driven SBARDEF format.